Add AddToCalendar component and replace usages
Introduce a reusable AddToCalendar component (app/components/add-to-calendar/add-to-calendar.component.tsx) that provides a unified UI for Google Calendar and Apple (.ics) actions and handles mobile vs desktop behavior. Replace ad-hoc calendar dropdowns and ICS handling in journey-finder-sign-up, set-a-reminder, and outreach-details components to use the new component, removing duplicated icon/svg, device-detection, and dropdown logic. Also wire up googleCalendarLink imports and simplify the outreach button to support an icon-only compact variant.
